+++ script/run.py
@@ -5,8 +5,8 @@
 import argparse
 import tkinter as tk
 import numpy as np
-from plan_config import PlanConfig
-from plan_viz import PlanViz
+from plan_config import PlanConfig2023, PlanConfig2024
+from plan_viz import PlanViz2023, PlanViz2024
 
 
 def main() -> None:
@@ -14,6 +14,7 @@ def main() -> None:
     """
     parser = argparse.ArgumentParser(description="Plan visualizer for a MAPF instance")
     parser.add_argument("--map", type=str, help="Path to the map file")
+    parser.add_argument("--year", type=str, help="Competition Year")
     parser.add_argument("--plan", type=str, help="Path to the planned path file")
     parser.add_argument("--n", dest="team_size", type=int, default=np.inf,
                         help="Number of agents")
@@ -41,12 +42,20 @@ def main() -> None:
     parser.add_argument("--heu", dest="heu_file", type=str, default="",
                         help="Show the low-level heuristics")
     args = parser.parse_args()
-
-    plan_config = PlanConfig(args.map, args.plan, args.team_size, args.start, args.end,
-                             args.ppm, args.moves, args.delay, args.heat_maps, args.hwy_file,
-                             args.search_tree_files, args.heu_file)
-    PlanViz(plan_config, args.show_grid, args.show_ag_idx, args.show_task_idx,
-            args.show_static, args.show_conf_ag)
+    if args.year == "2023":
+        plan_config = PlanConfig2023(args.map, args.plan, args.team_size, args.start, args.end,
+                                args.ppm, args.moves, args.delay, args.heat_maps, args.hwy_file,
+                                args.search_tree_files, args.heu_file)
+        PlanViz2023(plan_config, args.show_grid, args.show_ag_idx, args.show_task_idx,
+                args.show_static, args.show_conf_ag)
+    elif args.year == "2024":
+        plan_config = PlanConfig2024(args.map, args.plan, args.team_size, args.start, args.end,
+                              args.ppm, args.moves, args.delay)
+        PlanViz2024(plan_config, args.show_grid, args.show_ag_idx, args.show_task_idx,
+                args.show_static, args.show_conf_ag)
+    else:
+        print(f"Year wrong: {args.year} (should be 2023 or 2024)")
+        return
     tk.mainloop()
